Fixed deletion of Achievements

+++ HiP-Achievements/Controllers/AchievementsController.cs
@@ -150,18 +150,13 @@ namespace PaderbornUniversity.SILab.Hip.Achievements.Controllers
 
             if (!string.IsNullOrEmpty(achievement.Filename))
             {
-                System.IO.File.Delete(achievement.Filename);
+                if (System.IO.File.Exists(achievement.Filename))
+                    System.IO.File.Delete(achievement.Filename);
                 await _thumbnailService.TryClearThumbnailCacheAsync(id);
             }
 
-            var ev = new AchievementDeleted()
-            {
-                Id = id,
-                UserId = User.Identity.GetUserIdentity(),
-                Timestamp = DateTime.Now
-            };
-
-            await _eventStore.AppendEventAsync(ev);
+            //we use the BaseResourceType here
+            await EntityManager.DeleteEntityAsync(_eventStore, ResourceTypes.Achievement, id, User.Identity.GetUserIdentity());
             return NoContent();
         }
